Add files via upload
This is the app for MechaNurse Prototype. Since its a prototype so Its not completely built ground up by me its made in an IOT platform named Blynk. The app is given as a QR code. Download the Blynk app and scan the QR code there.
To start using it:
1. Download the blynk app.
2. Touch the QR code Icon and scan the QR code.